python casting for H5 and yaml in new file
